# Use a faster algorithm for palindromes