--- a/lib/filter.js
+++ b/lib/filter.js
@@ -11,9 +11,13 @@ exports.filterOnlyLetters = function(letters) {
 };
 
 exports.filterPalindromes = function() {
-  const isPalindrome = w => w
-  .split('')
-  .reverse()
-  .join('') === w;
+  const isPalindrome = w => {
+    var truncHalf = Math.floor(w.length / 2);
+    for(var i = 0; i < truncHalf; i++) {
+      let c1 = w[i], c2 = w[w.length - 1 -i];
+      if(c1 !== c2) return false;
+    }
+    return true;
+  };
   return sowpods.filter(isPalindrome);
 };