Measured in Node:

Slow (hr): 0s 215.737981ms
Fast (hr): 0s 9.183184ms
Slow (hr): 0s 156.637602ms
Fast (hr): 0s 6.659634ms
Slow (hr): 0s 148.845255ms
Fast (hr): 0s 6.122991ms
Slow (hr): 0s 155.119437ms
Fast (hr): 0s 6.105934ms
Slow (hr): 0s 154.708654ms
Fast (hr): 0s 6.322454ms